Upgrade everything to latest versions
This includes a bunch of small changes to adapt to how atrium-api has changed
over time. They're not functional or interesting, just some type-level
adjustments that are needed.

Some more complicated logic was changed in how profile details are parsed,
since atrium's way of doing things is weird and hard to understand so I just
manually grab stuff from the object map instead of relying on atrium's types.
This is similar to how CBOR parsing is done.

Boring maintenance stuff.

+++ src/algos.rs
@@ -19,7 +19,7 @@ pub trait Algo {
     async fn fetch_posts(
         &self,
         database: &Database,
-        limit: i32,
+        limit: u8,
         earlier_than: Option<(DateTime<Utc>, &str)>,
     ) -> Result<Vec<database::Post>>;
 }
