+++ src/algos/nederlandskie.rs
@@ -50,7 +50,7 @@ impl Algo for Nederlandskie {
     async fn fetch_posts(
         &self,
         database: &Database,
-        limit: i32,
+        limit: u8,
         earlier_than: Option<(DateTime<Utc>, &str)>,
     ) -> Result<Vec<database::Post>> {
         Ok(database
